Extract default extension helpers
