+++ .tickets/tlht-632w.md
@@ -1,6 +1,6 @@
 ---
 id: tlht-632w
-status: open
+status: closed
 deps: []
 links: []
 created: 2026-05-20T19:03:26Z
