+++ scripts/tlh-defaults.mjs
@@ -5,6 +5,13 @@ import { homedir } from "node:os";
 import { fileURLToPath } from "node:url";
 import process from "node:process";
 
+import {
+	disabledDefaultExtensionIds as disabledIdsFromSettings,
+	packageIdentity,
+	packageSourceOf,
+	readDefaultExtensions,
+} from "./lib/default-extensions.mjs";
+
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
@@ -115,111 +122,12 @@ function readJson(path, { missingValue } = {}) {
 	}
 }
 
-function readStringArrayField(entry, key, label) {
-	if (entry[key] === undefined) return [];
-	if (!Array.isArray(entry[key])) {
-		throw new Error(`Default extension ${label} field '${key}' must be an array`);
-	}
-	return entry[key]
-		.map((value) => (typeof value === "string" ? value.trim() : ""))
-		.filter((value) => value.length > 0);
-}
-
-function readBooleanField(entry, key, label) {
-	if (entry[key] === undefined) return false;
-	if (typeof entry[key] !== "boolean") {
-		throw new Error(`Default extension ${label} field '${key}' must be a boolean`);
-	}
-	return entry[key];
-}
-
-function readDefaultExtensions(path) {
-	const raw = readJson(path);
-	if (!Array.isArray(raw)) {
-		throw new Error(`Default extension manifest must be an array: ${path}`);
-	}
-
-	const seenIds = new Set();
-	const seenSources = new Set();
-	return raw.map((entry, index) => {
-		if (!isPlainObject(entry)) {
-			throw new Error(`Default extension entry ${index + 1} must be an object`);
-		}
-		const id = typeof entry.id === "string" ? entry.id.trim() : "";
-		const source = typeof entry.source === "string" ? entry.source.trim() : "";
-		const description = typeof entry.description === "string" ? entry.description.trim() : "";
-		const aliases = readStringArrayField(entry, "aliases", id || String(index + 1));
-		const replaces = readStringArrayField(entry, "replaces", id || String(index + 1));
-		const migrateReplacements = readBooleanField(entry, "migrateReplacements", id || String(index + 1));
-		const critical = readBooleanField(entry, "critical", id || String(index + 1));
-		if (!id) throw new Error(`Default extension entry ${index + 1} is missing id`);
-		if (!source) throw new Error(`Default extension ${id} is missing source`);
-		for (const candidateId of [id, ...aliases]) {
-			if (seenIds.has(candidateId)) throw new Error(`Duplicate default extension id or alias: ${candidateId}`);
-			seenIds.add(candidateId);
-		}
-		if (seenSources.has(packageIdentity(source))) throw new Error(`Duplicate default extension source: ${source}`);
-		seenSources.add(packageIdentity(source));
-		return { id, aliases, replaces, migrateReplacements, critical, source, description };
-	});
-}
-
 function isPlainObject(value) {
 	return Boolean(value) && typeof value === "object" && !Array.isArray(value);
 }
 
 function clone(value) {
 	return JSON.parse(JSON.stringify(value));
-}
-
-function packageSourceOf(entry) {
-	if (typeof entry === "string") return entry;
-	if (isPlainObject(entry) && typeof entry.source === "string") return entry.source;
-	return undefined;
-}
-
-function npmIdentity(spec) {
-	const withoutPrefix = spec.slice("npm:".length).trim();
-	if (!withoutPrefix) return spec;
-	if (withoutPrefix.startsWith("@")) {
-		const secondAt = withoutPrefix.indexOf("@", 1);
-		return `npm:${secondAt === -1 ? withoutPrefix : withoutPrefix.slice(0, secondAt)}`;
-	}
-	const firstAt = withoutPrefix.indexOf("@");
-	return `npm:${firstAt === -1 ? withoutPrefix : withoutPrefix.slice(0, firstAt)}`;
-}
-
-function gitIdentity(source) {
-	let value = source.trim();
-	if (value.startsWith("git:")) value = value.slice("git:".length).trim();
-	value = value.replace(/^https?:\/\//i, "");
-	value = value.replace(/^ssh:\/\//i, "");
-	value = value.replace(/^git:\/\//i, "");
-	value = value.replace(/^git@([^:]+):/, "$1/");
-	value = value.replace(/#.*$/, "");
-	value = value.replace(/\.git$/, "");
-	value = value.replace(/\.git(?=@)/, "");
-
-	const lastSlash = value.lastIndexOf("/");
-	const refAt = lastSlash === -1 ? -1 : value.indexOf("@", lastSlash + 1);
-	if (refAt !== -1) value = value.slice(0, refAt);
-
-	return `git:${value.toLowerCase()}`;
-}
-
-function packageIdentity(entry) {
-	const source = packageSourceOf(entry);
-	if (!source) return undefined;
-	const trimmed = source.trim();
-	if (trimmed.startsWith("npm:")) return npmIdentity(trimmed);
-	if (
-		trimmed.startsWith("git:") ||
-		/^(https?|ssh|git):\/\//i.test(trimmed) ||
-		trimmed.startsWith("git@")
-	) {
-		return gitIdentity(trimmed);
-	}
-	return `local:${trimmed}`;
 }
 
 function packageEntryDisablesExtensions(entry) {
@@ -232,38 +140,6 @@ function packageEntryDisablesExtensions(entry) {
 		.filter((value) => typeof value === "string")
 		.map((value) => value.trim())
 		.some((value) => disablingPatterns.has(value));
-}
-
-function rawDisabledIdsFromSettings(settings) {
-	if (!isPlainObject(settings)) return new Set();
-	const values = settings.tlh?.disabledDefaultExtensions;
-	if (!Array.isArray(values)) return new Set();
-	return new Set(values.filter((value) => typeof value === "string" && value.trim()).map((value) => value.trim()));
-}
-
-function criticalDefaultExtensionOptOutIds(defaultExtensions) {
-	const ids = new Set();
-	for (const extension of defaultExtensions) {
-		if (extension.critical !== true) continue;
-		ids.add(extension.id);
-		for (const alias of extension.aliases) ids.add(alias);
-	}
-	return ids;
-}
-
-function disabledIdsFromSettings(settings, defaultExtensions = []) {
-	const ids = rawDisabledIdsFromSettings(settings);
-	for (const id of criticalDefaultExtensionOptOutIds(defaultExtensions)) {
-		ids.delete(id);
-	}
-	for (const extension of defaultExtensions) {
-		if (extension.critical === true) continue;
-		if ([extension.id, ...extension.aliases].some((id) => ids.has(id))) {
-			ids.add(extension.id);
-			for (const alias of extension.aliases) ids.delete(alias);
-		}
-	}
-	return ids;
 }
 
 function validateSettings(settings) {
